Update discount data type from float to integer

+++ db/migrate/20220114004821_create_bulk_discounts.rb
@@ -1,7 +1,7 @@
 class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
-      t.float :discount
+      t.integer :discount
       t.integer :threshold
 
       t.timestamps
